Correctly decide if session involves a keyshare server
Previously this was decided by checking whether any of the scheme managers of any of the involved credetials and attributes that occur in the session uses a keyshare server. However, in case of disclosure and signing sessions, this could erroneously include attributes from disjunctions that have not been selected by the user.