IRMA server sessions are now started with instances of irma.RequestorRequest
That is, irma.ServiceProviderRequest, irma.SignatureRequestorRequest, or irma.IdentityProviderRequest.
These structs contain new fields with which the requestor can configure the session
(they have in fact always been present in session request JWTs). Functions and endpoints
that start an IRMA session now support one of these; or an instance of irma.SessionRequest;
or JSON-serialized versions of any of those.