fix: changed JSON representation of inner conjunctions to always keep track of order
Within a session request, inner conjunctions requesting specific attribute values
used to be serialized to a map. JSON and Go maps are however unordered, and the
order matters: if the requested attributes are received in a different order then
the session will fail. Inner conjunctions using values are now serialized using
the default Go serializer, resulting in a list of requested attributes with their values.