feat: allow disclosure of null attributes
Null attributes are attributes with bigint value 0, meaning that the issuer skipped that attribute. Instead of aborting the session, the irmaclient now allows disclosures of null attributes, when the verifier requests it. This allows the irmaclient to prove that a particular attribute was not issued to it, and allows verifiers to request optional attributes alongside other attributes.