More specific server responses
The server now responds with a "required attribute missing" error in case
a non-optional attribute is missing from the request, after trying
to redownload the configuration to see if it was tagged optional in the
meantime. This change required to extend the IrmaIdentifierSet with an
extra field that tracks missing AttributeTypeIdentifiers. Code that
populates this struct is also updated accordingly for possible future
consumers.