feat: support chained sessions
In the session request, the requestor can specify a nextSessionUrl.
After the session, the IRMA server POSTs the result to this URL, which
must respond with a next session request. The IRMA server then
automatically starts and performs the next session with the IRMA client.

This feature modifies the last message from the IRMA server to the client
in the protocol and thus introuduces a new protocol version. Both the
server and the client are backwards compatible with previous protocol
versions.