feat: add issuance wizards
This adds support for "issuance wizards" in requestor schemes.

The purpose of an issuance wizard is to guide IRMA app users through a chain of
issuance sessions, for example when the user is registering at a service that
requires the user to disclose attributes coming from multiple issuers. An
issuance wizard will generally mainly consist of credential types, which are
issued in the IRMA app to the user during execution of the wizard, but they can
also end with a disclosure session, or with sending the user to a website.
Issuance wizards can additionally contain text that explains to the user what is
going on.

This also introduces the notion of dependencies of credential types: a
credential type can now depend on one or more other credential types. If
credential type A depends on B and C, and credential type A is included in a
wizard, then the wizard will automatically also include B and C. (This can be
disabled in the wizard if desired, so that all required credential types must be
explicitly included in it.)