fix: move locking and hashing from middleware into datastore's get function as the session needs to be locked immediately and always (due to the timeout check in Redis' clientGet).
This makes the earlier introduced locking and updating middlewares obsolete.

An eventual update and unlock is however needed everywhere where the session first was retrieved. This has been extracted to a helper function now. This helper function will also get used in the library flows.

SubscribeServerSentEvents needed refactoring to avoid a deadlock as a consequence of locking directly in the datastore get function. Hereby, the exported function SubscribeServerSentEvents was removed from the codebase. This function should not have been used by library users anyway. Instead, the SessionStatus function should be used.