feat: support languages other than en, nl in schemes and scheme objects
Issuer schemes and requestor schemes can now specify the language of all content that they contain in new language XML tags or JSON arrays. During scheme parsing and validation, it is verified that all translatable items contained in the scheme are translated in the specified languages. If no languages are set in the scheme, then no such validation is performed.

Issuers and credential types can also specify their languages in the same way. If no language is set in the issuer, those of the containing scheme are taken. If no language is set in the credential type, then those of the containing issuer are taken (or those of the containing scheme). Similarly, requestors and wizard may override the languages of the containing requestor scheme.