fix: prevent deadlock when next session URL responds slowly
If the next session URL responds so slowly that a polling lands in the middle, then:
1) the poll HTTP handler locks the session store, and first tries to lock the session before unlocking the session store;
2) the /proofs HTTP handler has the session locked, and tries locking the session store when the next session URL finally responds.

But 1) still has the sessionstore lock, so 2) cannot complete and unlock the session lock, so 1) cannot complete and release the session store lock. Deadlock.

The fix is to make the unlocking of the memory session store in 1) immediate and not conditional on locking the session.