test: add new revocable credentialtype to testdata keyshare scheme
Previously, the test.test.email credentialtype was used both in revocation integration test, as well as serving as the KeyshareAttribute of the test scheme. It being revocable means that it is difficult to issue in test settings, so here we remove revocation support for test.test.email, while adding a new credential type that supports revocation for use in the revocation integration tests.