fix: chained sessions bug in irmaclient when first session involves keyshare server and second does not
If in a chain of sessions attributes are disclosed in a session, then they are implicitly (automatically and invisible to the user) always disclosed in all later sessions of the chain. The keyshare code of the irmaclient failed to take into account these implicit disclosures, thus missing that a keyshare server may be involved in these implicit disclosures even when none is in the non-implicit part of the current session. This means that if a session involves disclosure involving a keyshare server while a later session does not involve a keyshare server, then that session fails. This commit fixes this by properly taking the implicit disclosures into account in the keyshare code of the irmaclient.